Make bsdfetch the "default" fetch, remove conflicts
